src/pi/monitor/requirements.txt
gpsd-py3==0.3.0
pyserial==3.5
influxdb_client==1.47.0